generate_key func move util.py
