+++ util.py
@@ -1,3 +1,4 @@
+import random
 from io import BytesIO
 import zipfile
 
@@ -10,3 +11,13 @@ def generate_zip(files):
             zf.writestr(f[0], f[1])
 
     return mem_zip.getvalue()
+
+
+keys = random.sample(range(10000, 100000), 1000)
+key_num = 0
+
+
+def generate_key():
+    global key_num
+    key_num += 1
+    return keys[key_num]
